+++ lib/Statistics/Histogram.rb
@@ -25,10 +25,10 @@ module Statistics
 
     private
 
-    def initialize(values, bin_width: nil, bin_count: nil, method: :square_root)
+    def initialize(values, bin_width: nil, bin_count: nil, method: :square_root, factor: nil)
       raise ArgumentError, 'Values must not be empty' if values.empty?
       @values = values.map(&:to_f).sort
-      @bin_width = Bin.width(@values, bin_width: bin_width, bin_count: bin_count, method: method)
+      @bin_width = Bin.width(@values, bin_width: bin_width, bin_count: bin_count, method: method, factor: factor)
       @boundaries = Bin.boundaries(@values, bin_width: @bin_width)
       @bins = allocate_values
     end